+++ config/jsdoc.js
@@ -1,5 +1,8 @@
 'use strict';
 
+/**
+ * @fileoverview Defines configuration for documentation output
+ */
 module.exports = {
 	opts: {
 		destination: './docs',