+++ src/Core.js
@@ -1,3 +1,6 @@
+/**
+ * @fileoverview Initializes the Grindstone constructor and exports for consumption
+ */
 (function(root, lib) {
 	if (typeof exports !== 'undefined') {
 		return (module.exports = lib());