+++ src/utils/index.js
@@ -1,3 +1,6 @@
+/**
+ * @fileoverview Defines a set of reusable shared methods for use throughout the library
+ */
 const utils = (() => {
 	const applyContext = (item, selector, set) => {
 		const elems = item.querySelectorAll(selector);